requirement.txt
pandas
os
dotenv
requests
mysql.connector
pymysql
